Corrected getLayerString to be size_t[layer] long int[side]   This allows -ve value to be used to get complement   Two functions had side/layer swapped round so I    have added ELog::EM<<ELog::endErr; lines to remind you    that has happened and to change the calling code.

+++ Model/essBuild/ButterflyModerator.cxx
@@ -252,7 +252,7 @@ ButterflyModerator::getLayerSurf(const size_t,const long int) const
 }
 
 std::string
-ButterflyModerator::getLayerString(const size_t,const size_t) const
+ButterflyModerator::getLayerString(const size_t,const long int) const
   /*!
     Only components have reference values
   */
