+++ Model/essBuildInc/CylPreMod.h
@@ -99,8 +99,8 @@ class CylPreMod : public attachSystem::ContainedGroup,
   const std::shared_ptr<BlockAddition>& getBox(const char) const;
   std::string getBoxCut(const char) const;
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
 
 };
