+++ Model/essBuildInc/DiskPreMod.h
@@ -80,7 +80,7 @@ class DiskPreMod : public attachSystem::ContainedComp,
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const long int) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
   /// total height of object
   double getZOffset() const { return zStep; }
