+++ Model/t1UpgradeInc/CH4PreFlat.h
@@ -80,7 +80,7 @@ class CH4PreFlat : public CH4PreModBase
   virtual int 
     getLayerSurf(const size_t,const long int) const;
   virtual std::string 
-    getLayerString(const size_t,const size_t) const;
+    getLayerString(const size_t,const long int) const;
 
   
   virtual void createAll(Simulation&,const attachSystem::FixedComp&,
