+++ Model/t1UpgradeInc/CH4PreMod.h
@@ -83,7 +83,7 @@ class CH4PreMod : public CH4PreModBase
   virtual int 
     getLayerSurf(const size_t,const long int) const;
   virtual std::string 
-    getLayerString(const size_t,const size_t) const;
+    getLayerString(const size_t,const long int) const;
 
   void createAll(Simulation&,const attachSystem::FixedComp&,
 		 const size_t,const size_t);
